TST: Move to `utils.testing`
