+++ pysatModels/tests/test_utils_extract.py
@@ -154,67 +154,6 @@ class TestUtilsExtractInstThroughMod(object):
         assert str(err).find(err_msg) >= 0
         return
 
-    def test_failure_non_timelike_time(self):
-        """Test the failure when first dimension not timelike."""
-
-        if self.func != extract.extract_modelled_observations:
-
-            self.model['time'] = np.arange(0, len(self.model['time']))
-
-            # Run, raising a value error
-            with pytest.raises(ValueError) as err:
-                self.func(*self.input_args, **self.input_kwargs)
-
-            assert str(err).find('does not appear to be a time dimension') >= 0
-
-        return
-
-    def test_compare_model_name_coordinates_success(self):
-        """Ensure `compare_model_name_coordinates` works for proper inputs."""
-
-        compare_mod_name_coordinates(self.model['dummy1'],
-                                     ['latitude', 'longitude'])
-        return
-
-    @pytest.mark.parametrize("var,coords,msg,flag", [('dummy1',
-                                                      ['longitude', 'latitude'],
-                                                      'not in same order as',
-                                                      False),
-                                                     ('dummy1',
-                                                      ['wrong', 'latitude'],
-                                                      'not all within variable',
-                                                      False),
-                                                     ('dummy1',
-                                                      ['latitude', 'longitude'],
-                                                      'appear to be a time',
-                                                      True)])
-    def test_compare_model_name_coordinates_failure(self, var, coords, msg,
-                                                    flag):
-        """Ensure `compare_mod_name_coordinates` works for proper inputs.
-
-        Parameters
-        ----------
-        var : str
-            Variable to test.
-        coords : list
-            List of coordinate variable names.
-        msg : str
-            Failure message to test for.
-        flag : bool
-            If True, replace time index with integers.
-
-        """
-
-        if flag:
-            self.model['time'] = np.arange(0, len(self.model['time']))
-
-        with pytest.raises(ValueError) as err:
-            compare_mod_name_coordinates(self.model[var], coords)
-
-        assert str(err).find(msg) >= 0
-
-        return
-
     def test_failure_for_already_ran_data(self, caplog):
         """Test the failure for all model variables already extracted."""
 
@@ -864,71 +803,3 @@ class TestUtilsExtractInstModIrregViewXarray(TestUtilsExtractInstModIrregView):
         return
 
 
-def compare_mod_name_coordinates(data, mod_name):
-    """Compare ordering of mod_name against data coordinates, ignoring time.
-
-    First dimension for `data` needs to be the main DateTimeIndex.
-
-    Parameters
-    ----------
-    data : xarray.Dataseries
-        Variable from xarray to compare against `mod_name`.
-    mod_name : list
-        List of coordinate names.
-
-    Raises
-    ------
-    ValueError if `mod_name` and `data` aren't consistent.
-
-    """
-
-    # Get a list of all spatial coordinates, time is presumed to be first.
-    # Coordinates aren't in actual data ordering though :(
-    # Get the ordering of coordinates by using the ordering of data.dims.
-    # Collect dims for coordinates and then find the association with data.dims
-    # to construct correct order for data coordinates.
-    raw_coords = list(data.coords)
-    coord_dims = []
-    for coord in raw_coords:
-        coord_dims.append(data[coord].dims[0])
-
-    # Get list of dimension names
-    dims = list(data.dims)
-
-    # Check first dimension is full of datetime data
-    if str(data[data.dims[0]].data.dtype).find('datetime') < 0:
-        estr = ''.join(['First dimension of ', data.dims[0], ' does not ',
-                        'appear to be a time dimension.'])
-        raise ValueError(estr)
-    else:
-        # Drop time-like dimension
-        dims = dims[1:]
-
-    # Construct dim order for coordinates
-    coords = []
-    for dim in dims:
-        for i, coord in enumerate(coord_dims):
-            if dim == coord:
-                coords.append(raw_coords[i])
-
-    # Check actual coordinates all present in `mod_name`
-    missing_coords = []
-    for coord in coords:
-        if coord not in mod_name:
-            missing_coords.append(coord)
-
-    if len(missing_coords) > 0:
-        estr = ''.join(['Provided coordinates ', repr(mod_name), ' are not ',
-                        'all within variable coordinates ',
-                        repr(missing_coords), '.'])
-        raise ValueError(estr)
-
-    # Check order for both sets of coordinates
-    for i, name in enumerate(mod_name):
-        if name != coords[i]:
-            estr = ''.join(['Provided coordinates ', repr(mod_name),
-                            ' not in same order as variable coordinates ',
-                            repr(coords)])
-            raise ValueError(estr)
-
-    return
